refactor: remove components/forms/connector/source folder
