--- src/components/forms/connector/index.ts
@@ -1,1 +1,0 @@
-export * from "./source";
