--- src/components/forms/connector/source/ConfigureSourceForm/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./ConfigureSourceForm";
-export * from "./ConfigureSourceForm";
