--- src/components/forms/connector/source/CreateSourceForm/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./CreateSourceForm";
-export * from "./CreateSourceForm";
