+++ src/components/forms/index.ts
@@ -1,4 +1,3 @@
-export * from "./connector";
 export * from "./model";
 export * from "./pipeline";
 export * from "./commons";
